refactor: StringParser로 이동
