+++ 5-period/java-bridge-main/src/main/java/bridge/util/StringParser.java
@@ -2,8 +2,12 @@ package bridge.util;
 
 import bridge.exception.CustomIllegalArgumentException;
 import bridge.exception.ErrorMessage;
+import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 public class StringParser {
 
@@ -32,5 +36,19 @@ public class StringParser {
         return Arrays.stream(input.split(delimiter, CONTAINS_EMPTY))
                 .map(String::strip)
                 .toList();
+    }
+
+    /**
+     * 정규식 패턴으로 문자열을 파싱하여 매칭되는 그룹들의 리스트를 반환
+     */
+    public static List<String> findMatchingGroups(String input, Pattern pattern) {
+        Matcher matcher = pattern.matcher(input);
+        List<String> groups = new ArrayList<>();
+        while (matcher.find()) {
+            for (int i = 1; i <= matcher.groupCount(); i++) {
+                groups.add(matcher.group(i));
+            }
+        }
+        return Collections.unmodifiableList(groups);
     }
 }
